requirements.txt
ccxt==4.1.58
termcolor==2.3.0
python-dotenv==1.0.0
